+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 3.1.1
+
+* Bugfix
+  * `DaviModel` and scroll controllers being disposed by `Davi` generating an error that was used after being disposed.
+
 ## 3.1.0
 
 * Adding semantics on header and cells.